Add compilation options for windows. Partition platform specific aspects to platform specific modules. Minor re-arrangements. Add stub implementations for windows constructors: inputForConfig, outputForConfig.
These are only the changes required for the library to compile on windows. Nothing more. The tests, for instance, do not compile.
