+++ foreign/posix/gwinsz.h
