+++ foreign/posix/set_term_timing.c
