+++ foreign/vty_mk_wcwidth.c
