+++ src/Graphics/Vty/Input/Posix/Loop.hs
@@ -11,11 +11,12 @@
 -- behavior is still under tested. I should collect more of these examples...
 --
 -- reference: http://www.unixwiz.net/techtips/termios-vmin-vtime.html
-module Graphics.Vty.Input.Loop where
+module Graphics.Vty.Input.Posix.Loop where
 
 import Graphics.Vty.Config
 import Graphics.Vty.Input.Classify
 import Graphics.Vty.Input.Events
+import Graphics.Vty.Input.Interface
 
 import Control.Applicative
 import Control.Concurrent
@@ -40,21 +41,6 @@ import System.Posix.Terminal
 import System.Posix.Types (Fd(..))
 
 import Text.Printf (hPrintf)
-
-data Input = Input
-    { -- | Channel of events direct from input processing. Unlike 'nextEvent' this will not refresh
-      -- the display if the next event is an 'EvResize'.
-      _eventChannel  :: TChan Event
-      -- | Shuts down the input processing. This should return the terminal input state to before
-      -- the input initialized.
-    , shutdownInput :: IO ()
-      -- | Changes to this value are reflected after the next event.
-    , _configRef :: IORef Config
-      -- | input debug log
-    , _inputDebug :: Maybe Handle
-    }
-
-makeLenses ''Input
 
 data InputBuffer = InputBuffer
     { _ptr :: Ptr Word8
@@ -143,7 +129,7 @@ parseEvent = do
             logMsg $ "remaining: " ++ show remaining
             unprocessedBytes .= remaining
             return e
-        _                   -> mzero 
+        _                   -> mzero
 
 dropInvalid :: InputM ()
 dropInvalid = do
