+++ src/Graphics/Vty/Input/Terminfo/Classify.hs
@@ -1,11 +1,4 @@
 {-# OPTIONS_HADDOCK hide #-}
--- This makes a kind of tri. Has space efficiency issues with large input blocks.
--- Likely building a parser and just applying that would be better.
--- I did not write this so I might just rewrite it for better understanding. Which is not the best
--- of reasons.
--- TODO: measure and rewrite if required.
--- TODO: The ClassifyMap interface requires this code to always assure later entries override
--- earlier.
 module Graphics.Vty.Input.Classify
     ( classify
     , KClass(..)
@@ -47,7 +40,7 @@ compile table = cl' where
                 -- produced.
                 -- The test verifyFullSynInputToEvent2x verifies this.
                 -- H: There will always be one match. The prefixSet contains, by definition, all
-                -- prefixes of an event. 
+                -- prefixes of an event.
                 False ->
                     let inputPrefixes = reverse $ take maxValidInputLength $ tail $ inits inputBlock
                     in case mapMaybe (\s -> (,) s `fmap` M.lookup s eventForInput) inputPrefixes of
