+++ src/Graphics/Vty/Input/Windows.hs
@@ -1,0 +1,8 @@
+module Graphics.Vty.Input.Windows where
+
+import Graphics.Vty.Config
+import Graphics.Vty.Input.Events
+import Graphics.Vty.Input.Interface
+
+inputForConfig :: Config -> IO Input
+inputForConfig _ = undefined
