+++ src/Graphics/Vty/Output/Windows.hs
@@ -1,0 +1,7 @@
+module Graphics.Vty.Output.Windows where
+
+import Graphics.Vty.Config
+import Graphics.Vty.Output.Interface
+
+outputForConfig :: Config -> IO Output
+outputForConfig _ = undefined
